Revert "headless: refact request as global var (#319)"
This reverts commit c625da93e39b2e3c20de7921c3518494ea18fe2c.
