+++ Products/zms/_objattrs.py
@@ -36,7 +36,7 @@ from Products.zms import _globals
 #  getobjattrdefault
 # ------------------------------------------------------------------------------
 def getobjattrdefault(obj, obj_attr, lang):
-    request = obj.get('REQUEST', _globals.headless_http_request)
+    request = obj.get('REQUEST', standard.create_headless_http_request())
     v = None
     datatype = obj_attr['datatype_key']
     default = None
@@ -595,7 +595,7 @@ class ObjAttrs(object):
     #  attr({key0:value0,...,keyN:valueN}) -> setObjProperty(key0,value0),...
     # --------------------------------------------------------------------------
     def attr(self, *args, **kwargs):
-      req = self.get('REQUEST', _globals.headless_http_request)
+      req = self.get('REQUEST', standard.create_headless_http_request())
       request = kwargs.get('request',kwargs.get('REQUEST',req))
       if len(args) == 1 and isinstance(args[0], str):
         return self.getObjProperty( args[0], request, kwargs)
@@ -610,7 +610,7 @@ class ObjAttrs(object):
     #  ObjAttrs.evalMetaobjAttr
     # --------------------------------------------------------------------------
     def evalMetaobjAttr(self, *args, **kwargs):
-      request = self.get('REQUEST', _globals.headless_http_request)
+      request = self.get('REQUEST', standard.create_headless_http_request())
       root = self
       key = args[0]
       id = standard.nvl(request.get('ZMS_INSERT'), self.meta_id)
